+++ NEWS.md
@@ -1,4 +1,6 @@
 # arkhe 1.1.0.9000
+## New classes and methods
+* Add `math_lcm()` and `math_gcd()` to compute the lowest common multiple and the greatest common divisor.
 
 # arkhe 1.1.0
 ## New classes and methods